Remove PLAIN_TEXT and support ignoring other types while scanning
The PLAIN_TEXT is not supported by Axon.ivy Designer. The export
commands currenly stops if it detect the co.meta is not a TEXT or
STRING. But other types (SOURCE, etc) can also have similar structure so
checking solely based on the presense of *.data files is fragile.

Fix by returning None and the export command has to filter out None
values.